+++ lib/reference-panel/references-view.tsx
@@ -1,4 +1,10 @@
-import { CompositeDisposable, TextEditor } from 'atom';
+import {
+  CompositeDisposable,
+  DisplayMarker,
+  Emitter,
+  TextBuffer,
+  TextEditor
+} from 'atom';
 import { Minimatch } from 'minimatch';
 import etch from 'etch';
 import Path from 'path';
@@ -9,6 +15,7 @@ import * as console from '../console';
 
 import type { Reference } from 'atom-ide-base';
 import type { EtchComponent } from 'etch';
+import FindReferencesManager from '../find-references-manager';
 
 function isEtchComponent(el: unknown): el is EtchComponent {
   if (!el) return false;
@@ -16,16 +23,26 @@ function isEtchComponent(el: unknown): el is EtchComponent {
   return ('refs' in el) && ('element' in el);
 }
 
-function pluralize(singular: string, plural: string, count: number) {
-  return count > 1 ? plural : singular;
+
+function pluralize(count: number, singular: string, plural: string = `${singular}s`) {
+  let noun = count === 1 ? singular : plural;
+  return `${count} ${noun}`;
 }
 
 function describeReferences(referenceCount: number, fileCount: number, symbolName: string) {
   return (
     <span ref="previewCount" className="preview-count inline-block">
-      {referenceCount} {pluralize('result', 'results', referenceCount)} found in {fileCount} {pluralize('file', 'files', fileCount)} for {' '}
+      {pluralize(referenceCount, 'result')} found in {' '}
+      {pluralize(fileCount, 'file')} for {' '}
       <span className="highlight-info">{symbolName}</span>
     </span>
+  );
+}
+
+function descendsFrom(filePath: string, projectPath: string) {
+  if (typeof filePath !== 'string') return false;
+  return filePath.startsWith(
+    projectPath.endsWith(Path.sep) ? projectPath : `${projectPath}${Path.sep}`
   );
 }
 
@@ -34,13 +51,6 @@ function descendsFromAny(filePath: string, projectPaths: string[]): string | fal
     if (descendsFrom(filePath, projectPath)) return projectPath;
   }
   return false;
-}
-
-function descendsFrom(filePath: string, projectPath: string) {
-  if (typeof filePath !== 'string') return false;
-  return filePath.startsWith(
-    projectPath.endsWith(Path.sep) ? projectPath : `${projectPath}${Path.sep}`
-  );
 }
 
 function matchesIgnoredNames(filePath: string, ignoredNames: Minimatch[]) {
@@ -55,16 +65,15 @@ function matchesIgnoredNames(filePath: string, ignoredNames: Minimatch[]) {
 type SplitDirection = 'left' | 'right' | 'up' | 'down' | 'none';
 type FormalSplitDirection = 'left' | 'right' | 'up' | 'down' | undefined;
 
-type ReferencesViewProperties = {
-  ref?: string,
+type ReferencesViewContext = {
+  manager: FindReferencesManager,
+  editor: TextEditor,
+  marker: DisplayMarker,
   references: Reference[],
   symbolName: string
-};
+}
 
-let lastReferences: { references: Reference[], symbolName: string } = {
-  references: [],
-  symbolName: ''
-};
+type ReferencesViewProperties = { ref?: string } & ReferencesViewContext;
 
 function getOppositeSplit(split: SplitDirection): FormalSplitDirection {
   return {
@@ -76,45 +85,99 @@ function getOppositeSplit(split: SplitDirection): FormalSplitDirection {
   }[split] as FormalSplitDirection;
 }
 
+let panelId = 1;
+
 export default class ReferencesView {
+  // Base URI. We add `/1`, `/2`, etc., so that different instances of the
+  // panel can be distinguished.
   static URI = "atom://pulsar-find-references/results";
 
-  static instances: Set<ReferencesView> = new Set();
+  // Initialization data for panels that have not yet been instantiated.
+  static CONTEXTS: Map<string, ReferencesViewContext> = new Map();
 
-  static setReferences(references: Reference[], symbolName: string) {
-    console.log('ReferencesPaneView.setReferences:', references);
-    lastReferences = { references, symbolName };
+  // Instances of `ReferencesView`.
+  static instances: Map<string, ReferencesView> = new Map();
 
-    for (let instance of ReferencesView.instances) {
-      instance.update(lastReferences);
+  static nextUri () {
+    return `${ReferencesView.URI}/${panelId++}`;
+  }
+
+  static setReferences(
+    uri: string,
+    context: ReferencesViewContext
+  ) {
+    if (ReferencesView.instances.has(uri)) {
+      // This instance already exists, so we can update it directly.
+      ReferencesView.instances.get(uri)!.update(context);
+    } else {
+      // This instance will soon exist, so we'll store this data for future
+      // lookup.
+      ReferencesView.CONTEXTS.set(uri, context);
     }
   }
 
+  public uri: string;
+
+  // Whether this panel can be reused the next time the “Show Panel” command is
+  // invoked.
+  public overridable: boolean = true;
+
   private subscriptions: CompositeDisposable = new CompositeDisposable();
+
+  // Component properties.
   private references: Reference[];
   private symbolName: string;
+  private editor: TextEditor;
+  private marker: DisplayMarker;
+  private manager: FindReferencesManager;
+
   private ignoredNameMatchers: Minimatch[] | null = null;
   private splitDirection: SplitDirection = 'none';
 
+  private emitter: Emitter = new Emitter();
+
   private filteredAndGroupedReferences!: Map<string, Reference[]>;
 
+  // URIs of buffers in the current result set.
+  private uris: Set<string> = new Set();
+
+  // Keeps track of which result has keyboard focus.
   private activeNavigationIndex: number = -1;
   private lastNavigationIndex: number = -1;
 
-  private collapsedIndices: Set<number> = new Set();
+  private bufferCache: Map<string, TextBuffer> = new Map();
+  private indexToReferenceMap: Map<number, Reference> = new Map();
 
-  private pinned: boolean = false;
+  // Keeps track of which result groups are collapsed.
+  private collapsedIndices: Set<number> = new Set();
 
   private previewStyle: { fontFamily: string } = { fontFamily: '' };
 
   public element!: HTMLElement;
   public refs!: { [key: string]: HTMLElement };
 
-  constructor() {
-    ReferencesView.instances.add(this);
-    this.references = lastReferences.references;
-    this.symbolName = lastReferences.symbolName;
-    console.debug('ReferencesView constructor:', this.references, this.symbolName);
+  constructor(uri: string, props?: ReferencesViewContext) {
+    ReferencesView.instances.set(uri, this);
+    this.uri = uri;
+    let context: ReferencesViewContext
+    if (props) {
+      context = props;
+    } else if (ReferencesView.CONTEXTS.has(uri)) {
+      context = ReferencesView.CONTEXTS.get(uri)!
+    } else {
+      throw new Error(`Expected context data for URI: ${uri}`)
+    }
+
+    let { references, symbolName, editor, marker, manager } = context;
+    this.references = references;
+    this.symbolName = symbolName;
+    this.editor = editor;
+    this.marker = marker;
+    this.manager = manager;
+
+    ReferencesView.CONTEXTS.delete(uri);
+
+    console.debug('ReferencesView constructor:', this.uri, this.props);
 
     if (!this.references) {
       throw new Error(`No references!`);
@@ -124,12 +187,32 @@ export default class ReferencesView {
 
     etch.initialize(this);
 
-    this.element.addEventListener('mousedown', this.handleClick.bind(this));
-
     this.subscriptions.add(
       atom.config.observe('editor.fontFamily', this.fontFamilyChanged.bind(this)),
       atom.config.observe('core.ignoredNames', this.ignoredNamesChanged.bind(this)),
-      atom.config.observe('pulsar-find-references.panel.splitDirection', this.splitDirectionChanged.bind(this))
+      atom.config.observe('pulsar-find-references.panel.splitDirection', this.splitDirectionChanged.bind(this)),
+
+      atom.workspace.observeTextEditors((editor) => {
+        // Since this panel updates in real time, we should arguably fetch new
+        // references whenever _any_ editor changes. For now, we'll refetch
+        // whenever one of the files in the result set is edited, even though
+        // this could end up missing new references as they are created.
+        editor.onDidStopChanging((_event) => {
+          if (this.referencesIncludeBuffer(editor.getBuffer())) {
+            this.refreshPanel();
+          }
+        })
+      }),
+
+      // If the marker is destroyed or made invalid, it means a buffer change
+      // has caused us not to be able to track the logical position of the
+      // point that initially trigged this panel. This makes it impossible for
+      // us to continue to update the results, so the panel must close.
+      this.marker.onDidChange(() => {
+        if (this.marker?.isValid()) return;
+        this.close();
+      }),
+      this.marker.onDidDestroy(() => this.close())
     );
 
     atom.commands.add<Node>(
@@ -145,10 +228,14 @@ export default class ReferencesView {
         'core:move-to-bottom': this.moveToBottom.bind(this),
         'core:confirm': this.confirmResult.bind(this),
         'core:copy': this.copyResult.bind(this),
+        // Piggyback on the user's keybindings for these functions, since the
+        // UI is practically identical to that of `find-and-replace`.
         'find-and-replace:copy-path': this.copyPath.bind(this),
         'find-and-replace:open-in-new-tab': this.openInNewTab.bind(this),
       }
     );
+
+    this.element.addEventListener('mousedown', this.handleClick.bind(this));
 
     this.refs.pinReferences.addEventListener(
       'click',
@@ -156,24 +243,40 @@ export default class ReferencesView {
     );
 
     this.focus();
+
+    this.buildBufferCache()
+      .then((cache) => {
+        this.bufferCache = cache
+        return etch.update(this);
+      });
   }
 
+  // Pane items that provide `onDidChangeTitle` can trigger updates to their
+  // tab and window titles.
+  onDidChangeTitle (callback: () => void) {
+    return this.emitter.on('did-change-title', callback);
+  }
+
+  // Move keyboard focus to the previous visible result.
   moveUp() {
-    if (this.activeNavigationIndex === this.lastNavigationIndex) return;
+    if (this.activeNavigationIndex === 0) return;
     let index = this.findVisibleNavigationIndex(-1);
     if (index === null) return;
     this.activeNavigationIndex = index;
-    etch.update(this);
+    etch.update(this).then(() => this.ensureSelectedItemInView());
   }
 
+  // Move keyboard focus to the next visible result.
   moveDown() {
     if (this.activeNavigationIndex === this.lastNavigationIndex) return;
     let index = this.findVisibleNavigationIndex(1);
     if (index === null) return;
     this.activeNavigationIndex = index;
-    etch.update(this);
+    etch.update(this).then(() => this.ensureSelectedItemInView());
   }
 
+  // Move the navigation index some number of increments, skipping any results
+  // that are collapsed.
   findVisibleNavigationIndex(delta: number) {
     let current = this.activeNavigationIndex;
     while (true) {
@@ -252,7 +355,7 @@ export default class ReferencesView {
     let index = this.findElementIndexNearHeight(currentOffset - this.refs.scrollContainer.offsetHeight);
 
     this.activeNavigationIndex = index;
-    etch.update(this);
+    etch.update(this).then(() => this.ensureSelectedItemInView());
   }
 
   pageDown() {
@@ -262,38 +365,101 @@ export default class ReferencesView {
     let index = this.findElementIndexNearHeight(currentOffset + this.refs.scrollContainer.offsetHeight);
 
     this.activeNavigationIndex = index;
-    etch.update(this);
+    etch.update(this).then(() => this.ensureSelectedItemInView());
   }
 
   moveToTop() {
     this.activeNavigationIndex = 0;
-    etch.update(this);
+    etch.update(this).then(() => this.ensureSelectedItemInView());
   }
 
   moveToBottom() {
     this.activeNavigationIndex = this.lastNavigationIndex;
-    etch.update(this);
+    etch.update(this).then(() => this.ensureSelectedItemInView());
+  }
+
+  ensureSelectedItemInView() {
+    if (!this.activeElement) return;
+    let containerRect = this.refs.scrollContainer.getBoundingClientRect();
+    let itemRect = this.activeElement.getBoundingClientRect();
+
+    let delta: number;
+    if (itemRect.top < containerRect.top) {
+      delta = itemRect.top - containerRect.top;
+    } else if (itemRect.bottom > containerRect.bottom) {
+      delta = itemRect.bottom - containerRect.bottom;
+    } else {
+      return;
+    }
+    this.refs.scrollContainer.scrollTop += delta;
   }
 
   confirmResult() {
-    let element = this.activeElement;
-    if (!element) return;
-    let { filePath = '', lineNumberStr = '-1', range = '' } = element.dataset;
-    let lineNumber = Number(lineNumberStr);
+    if (!this.activeElement) return;
+    let metadata = this.getMetadataForTarget(this.activeElement);
+    if (!metadata) return;
 
-    this.openResult(filePath, lineNumber, range);
+    let { filePath, lineNumber, rangeSpec } = metadata;
+    this.openResult(filePath, lineNumber, rangeSpec);
   }
 
+  // Copy the line of text from the reference. (Of limited utility, but
+  // implemented for feature equivalence with the `find-and-replace` panel.)
   copyResult() {
-    // TODO
+    if (!this.activeElement) return;
+
+    let reference = this.indexToReferenceMap.get(this.activeNavigationIndex);
+    if (!reference) return;
+
+    if (!this.bufferCache.has(reference.uri)) return;
+
+    // All the buffers for results should be present in this cache because we
+    // preloaded them during render.
+    let buffer = this.bufferCache.get(reference.uri);
+    if (!buffer) return;
+
+    let text = buffer.lineForRow(reference.range.start.row);
+    if (!text) return;
+
+    atom.clipboard.write(text);
   }
 
+  // Copy the relative file path of the keyboard-focused reference.
+  // (Implemented for feature equivalence with the `find-and-replace` panel.)
   copyPath() {
-    // TODO
+    if (!this.activeElement) return;
+    const { filePath = null } = this.activeElement.dataset;
+    if (!filePath) return;
+    let [projectPath, relativePath] = atom.project.relativizePath(filePath);
+    if (projectPath && atom.project.getDirectories().length > 1) {
+      relativePath = Path.join(Path.basename(projectPath), relativePath);
+    }
+    atom.clipboard.write(relativePath);
   }
 
-  openInNewTab() {
-    // TODO
+  // Open the result in a new tab whether or not it already exists in the
+  // workspace.
+  async openInNewTab() {
+    if (!this.activeElement) return;
+
+    let metadata = this.getMetadataForTarget(this.activeElement);
+    if (!metadata) return;
+
+    let { filePath, lineNumber: row, rangeSpec } = metadata;
+    if (!filePath) return;
+
+    let editor;
+    let exists = atom.workspace.getTextEditors().filter(e => e.getPath() === filePath);
+    if (!exists) {
+      editor = await atom.workspace.open(
+        filePath,
+        { activatePane: false, activateItem: false }
+      ) as TextEditor;
+    } else {
+      editor = await atom.workspace.open(filePath) as TextEditor;
+    }
+
+    this.revealReferenceInEditor(filePath, row, rangeSpec, editor);
   }
 
   getElementAtIndex(index: number): HTMLElement | null  {
@@ -301,36 +467,63 @@ export default class ReferencesView {
     return element ? (element as HTMLElement) : null;
   }
 
+  // The element that has keyboard focus.
   get activeElement(): HTMLElement | null {
     if (this.activeNavigationIndex < 0) return null;
     return this.getElementAtIndex(this.activeNavigationIndex);
   }
 
-  async update({ references, symbolName }: ReferencesViewProperties) {
-    // Ignore new references when pinned.
-    if (this.pinned) return Promise.resolve();
-
+  async update({ references, symbolName, editor, marker, manager }: Partial<ReferencesViewProperties>) {
     let changed = false;
-    if (references.length === 0 && symbolName === '')
-      return Promise.resolve();
 
-    if (this.references !== references) {
+    if (references && this.references !== references) {
       this.references = references;
       this.filterAndGroupReferences();
+      this.indexToReferenceMap.clear();
+      this.bufferCache = await this.buildBufferCache();
+      changed = true;
+    }
+    if (symbolName && this.symbolName !== symbolName) {
+      this.symbolName = symbolName;
+      // Triggers an update of the tab title.
+      this.emitter.emit('did-change-title');
       changed = true;
     }
 
-    if (this.symbolName !== symbolName) {
-      this.symbolName = symbolName;
-      changed = true;
+    // These properties don't trigger re-renders, but they must still be
+    // updated if changed.
+    if (editor) {
+      this.editor = editor;
+    }
+    if (marker) {
+      this.marker = marker;
+    }
+    if (manager) {
+      this.manager = manager;
     }
 
     return changed ? etch.update(this) : Promise.resolve();
   }
 
   destroy() {
-    ReferencesView.instances.delete(this);
+    ReferencesView.instances.delete(this.uri);
     this.subscriptions.dispose();
+  }
+
+  // Close this window.
+  close() {
+    this.destroy();
+    const pane = atom.workspace.paneForItem(this);
+    if (!pane) return;
+    pane.destroyItem(this);
+  }
+
+  // Given a buffer, returns whether the buffer's file path matches any of the
+  // current references.
+  referencesIncludeBuffer (buffer: TextBuffer) {
+    let bufferPath = buffer.getPath()
+    if (!bufferPath) return false
+    return this.uris.has(bufferPath)
   }
 
   fontFamilyChanged(fontFamily: string) {
@@ -346,6 +539,17 @@ export default class ReferencesView {
     this.splitDirection = splitDirection;
   }
 
+  getMetadataForTarget (target: HTMLElement) {
+    if (!target.matches('[data-line-number][data-file-path]')) return null;
+    let {
+      filePath = '',
+      lineNumber: lineNumberString = '-1',
+      rangeSpec = ''
+    } = target.dataset;
+    let lineNumber = Number(lineNumberString);
+    return { filePath, lineNumber, rangeSpec };
+  }
+
   handleClick(event: MouseEvent) {
     if (!event.target) return;
     let target = (event.target as HTMLElement)?.closest('[data-navigation-index]') as HTMLElement;
@@ -354,16 +558,18 @@ export default class ReferencesView {
       let viewportXOffset = event.clientX;
       let targetRect = target.getBoundingClientRect();
 
+      // A bit of a hack, but copies the approach of the equivalent
+      // `find-and-replace` result handler. Distinguishes between a click on
+      // the result and a click on the disclosure triangle that
+      // collapses/expands results.
       if (target.matches('.list-item') && viewportXOffset - targetRect.left <= 16) {
         this.toggleResult(navigationIndex);
         return;
       }
 
-      if (target.matches('[data-line-number][data-file-path]')) {
-        let filePath = target.dataset.filePath ?? '';
-        let lineNumber = Number(target.dataset.lineNumber || '-1');
-        let rangeSpec = target.dataset.range ?? '';
-
+      let metadata = this.getMetadataForTarget(target);
+      if (metadata) {
+        let { filePath, lineNumber, rangeSpec } = metadata;
         this.openResult(filePath, lineNumber, rangeSpec);
       }
 
@@ -374,7 +580,6 @@ export default class ReferencesView {
 
     etch.update(this);
     event.preventDefault();
-    // this.activate();
   }
 
   activate(): Promise<void> {
@@ -387,25 +592,19 @@ export default class ReferencesView {
   }
 
   handlePinReferencesClicked() {
-    this.pinned = !this.pinned;
+    this.overridable = !this.overridable;
     etch.update(this);
   }
 
+  // Brings the user to the given reference on click.
   async openResult(
     filePath: string,
     row: number,
     rangeSpec: string,
     { pending = true }: { pending: boolean } = { pending: true }
   ) {
-    let referencesForFilePath = this.filteredAndGroupedReferences.get(filePath);
-    if (!referencesForFilePath) return;
-    let referencesForLineNumber = referencesForFilePath.filter(({ range }) => {
-      return range.start.row == row;
-    });
-    let ranges = referencesForLineNumber.map(r => r.range);
-    let targetRange = rangeSpec === '' ? ranges[0] : ranges.find(r => {
-      return r.toString() === rangeSpec;
-    });
+    // Find an existing editor in the workspace for this file or else create
+    // one if needed.
     let editor = await atom.workspace.open(
       filePath,
       {
@@ -414,23 +613,52 @@ export default class ReferencesView {
         split: getOppositeSplit(this.splitDirection)
       }
     ) as TextEditor;
+
+    this.revealReferenceInEditor(filePath, row, rangeSpec, editor);
+  }
+
+  revealReferenceInEditor(filePath: string, row: number, rangeSpec: string, editor: TextEditor) {
+    let referencesForFilePath = this.filteredAndGroupedReferences.get(filePath);
+    if (!referencesForFilePath) return
+
+    let referencesForLineNumber = referencesForFilePath.filter(({ range }) => {
+      return range.start.row == row;
+    });
+
+    let ranges = referencesForLineNumber.map(r => r.range);
+    let targetRange = rangeSpec === '' ? ranges[0] : ranges.find(r => {
+      return r.toString() === rangeSpec;
+    });
+
+    // Reveal the row the result is on if it happens to be folded.
     editor.unfoldBufferRow(row);
+
     if (ranges.length > 0) {
       // @ts-expect-error undocumented option
       editor.getLastSelection().setBufferRange(targetRange ?? ranges[0], { flash: true });
     }
+
     editor.scrollToCursorPosition();
   }
 
+  // Groups the references according to the files they belong to.
   filterAndGroupReferences(): Map<string, Reference[]> {
     let paths = atom.project.getPaths();
     let results = new Map<string, Reference[]>();
+    let uris = new Set<string>();
+
     if (!this.references) return results;
 
+    // Group references by file.
     for (let reference of this.references) {
       let { uri } = reference;
+      uris.add(uri);
       let projectPath = descendsFromAny(uri, paths);
+
+      // Ignore any results that aren't within this project.
       if (projectPath === false) continue;
+
+      // Ignore any results within ignored files.
       if (matchesIgnoredNames(uri, this.ignoredNameMatchers ?? [])) continue;
 
       let [_, relativePath] = atom.project.relativizePath(uri);
@@ -439,17 +667,22 @@ export default class ReferencesView {
         resultsForPath = [];
         results.set(relativePath, resultsForPath);
       }
+
       resultsForPath.push(reference);
     }
 
     this.filteredAndGroupedReferences = results;
+    this.uris = uris;
     return results;
   }
 
   get props(): ReferencesViewProperties {
     return {
       references: this.references ?? [],
-      symbolName: this.symbolName ?? ''
+      symbolName: this.symbolName ?? '',
+      editor: this.editor,
+      marker: this.marker,
+      manager: this.manager
     };
   }
 
@@ -463,11 +696,13 @@ export default class ReferencesView {
   }
 
   copy() {
-    return new ReferencesView();
+    let newUri = ReferencesView.nextUri();
+    return new ReferencesView(newUri, this.props);
   }
 
   getTitle() {
-    return 'Find References Results';
+    let { symbolName } = this;
+    return `“${symbolName}”: Find References Results`;
   }
 
   getIconName() {
@@ -484,8 +719,59 @@ export default class ReferencesView {
     referencesView.element.focus();
   }
 
+  // Assembles a map between reference URIs and `TextBuffer`s for child views
+  // to consult.
+  async buildBufferCache() {
+    let map = new Map<string, TextBuffer>();
+    let editors = atom.workspace.getTextEditors();
+    for (let editor of editors) {
+      let path = editor.getPath();
+      let buffer = editor.getBuffer();
+      if (path === undefined) continue;
+      if (map.has(path)) continue;
+      map.set(path, buffer);
+    }
+    // Any buffers that aren't present already in the work space can be created
+    // from files on disk.
+    for (let uri of this.uris) {
+      if (map.has(uri)) continue;
+      map.set(uri, await TextBuffer.load(uri));
+    }
+    return map;
+  }
+
+  // How do we keep refreshing the references panel as we make changes in the
+  // project?
+  //
+  // * Remember the cursor position that triggered the panel. Create a marker
+  //   to track the logical buffer position through edits.
+  // * Open the panel and show the results.
+  // * When you open the panel, add an `onDidStopChanging` observer to every
+  //   `TextEditor` in the project. The callback should return early if the
+  //   editor isn't changing a buffer that is in the result set; otherwise it
+  //   should re-request the list of references.
+  // * When references are re-requested, they should use the current buffer
+  //   position of the marker we created in step 1.
+  //
+  // This works for as long as the cursor position can be logically tracked. If
+  // the marker is invalidated, that means a change has completely surrounded
+  // it, and we can no longer affirm it refers to the same symbol. At this
+  // point, we close the panel.
+  async refreshPanel() {
+    if (!this.manager || !this.editor || !this.marker) return;
+    let bundle = await this.manager.findReferencesForProjectAtPosition(
+      this.editor,
+      this.marker.getBufferRange().start
+    )
+    if (!bundle || bundle.type === 'error') return;
+
+    await this.update({
+      references: bundle.references,
+      symbolName: bundle.referencedSymbolName
+    });
+  }
+
   render() {
-    // console.log('ReferencesView render:', this.props, this.activeNavigationIndex);
     let listStyle = {
       position: 'absolute',
       overflow: 'hidden',
@@ -494,26 +780,26 @@ export default class ReferencesView {
       right: '0'
     };
 
-    let index = this.filteredAndGroupedReferences;
     let children = [];
 
     let navigationIndex = 0;
-    for (let [relativePath, references] of index) {
+    for (let [relativePath, references] of this.filteredAndGroupedReferences) {
       let view = (
         <ReferenceGroupView
           relativePath={relativePath}
           references={references}
           navigationIndex={navigationIndex}
+          indexToReferenceMap={this.indexToReferenceMap}
           activeNavigationIndex={this.activeNavigationIndex}
+          bufferCache={this.bufferCache}
           isCollapsed={this.collapsedIndices.has(navigationIndex)}
         />
       );
-      // console.log('ReferencesView adding child:', view);
       children.push(view);
       navigationIndex += references.length + 1;
     }
 
-    this.lastNavigationIndex = navigationIndex;
+    this.lastNavigationIndex = navigationIndex - 1;
 
     let containerStyle =  {
       position: 'relative',
@@ -524,13 +810,16 @@ export default class ReferencesView {
     let matchCount = this.references.length;
     let classNames = cx('find-references-pane', 'preview-pane', 'pane-item', { 'no-results': matchCount === 0 });
 
-    let pinButtonClassNames = cx('btn', 'icon', 'icon-pin', { 'selected': this.pinned });
+    let pinButtonClassNames = cx('btn', 'icon', 'icon-pin', {
+      'selected': !this.overridable
+    });
+
     return (
       <div className={classNames} tabIndex={-1}>
         <div className="preview-header">
           {describeReferences(this.references.length, this.filteredAndGroupedReferences.size, this.symbolName)}
 
-          <div ref="pinReferences" className={pinButtonClassNames}>Pin references</div>
+          <div ref="pinReferences" className={pinButtonClassNames}>Don’t override</div>
         </div>
 
         <div ref="referencesView" className="results-view focusable-panel" tabIndex={-1} style={this.previewStyle}>